refactor(viewer): collapse five copies of get_results_dir into paths.py
Four of the five were byte-identical; ai_comparer's had drifted, accepting a
non-directory candidate and carrying an extra /evalbench/results entry that the
repo-root candidate already resolves to inside the container.
