+++ viewer/main.py
@@ -12,6 +12,7 @@ import precompute_trends
 import dataset_quality
 from summarizer import summarize_eval_scoring
 from ai_comparer import compare_evals
+from paths import get_results_dir
 
 @me.stateclass
 class State:
@@ -114,26 +115,6 @@ def df_to_config(df: pd.DataFrame) -> dict:
 
 
 
-
-
-def get_results_dir():
-    # Try to read from environment variable
-    res_dir = os.environ.get("RESULTS_DIR")
-    if res_dir:
-        return res_dir
-        
-    # Check multiple locations for results directory
-    results_dir_candidates = [
-        "/tmp_session_files/results",
-        os.path.join(os.path.dirname(os.path.dirname(__file__)), "results"),
-        os.path.join(os.getcwd(), "results"),
-    ]
-
-    for candidate in results_dir_candidates:
-        if os.path.exists(candidate) and os.path.isdir(candidate):
-            return candidate
-
-    return results_dir_candidates[1]  # Fallback to default
 
 
 # (mtime, value). Each is replaced as a whole so a concurrent reader never sees a
