+++ viewer/precompute_dataset_quality.py
@@ -13,6 +13,8 @@ import json
 import logging
 import os
 
+from paths import get_results_dir
+
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 CACHE_FILENAME = "dataset_quality_cache.json"
@@ -29,26 +31,6 @@ SUMMARY_COMPARATOR = "dataset_quality"
 # Category JSON blobs embed per-CUJ evidence and prose, and scores.csv rows are
 # correspondingly huge.
 csv.field_size_limit(10**9)
-
-
-def get_results_dir():
-    # Try to read from environment variable
-    res_dir = os.environ.get("RESULTS_DIR")
-    if res_dir:
-        return res_dir
-
-    # Check multiple locations for results directory
-    results_dir_candidates = [
-        "/tmp_session_files/results",
-        os.path.join(os.path.dirname(os.path.dirname(__file__)), "results"),
-        os.path.join(os.getcwd(), "results"),
-    ]
-
-    for candidate in results_dir_candidates:
-        if os.path.exists(candidate) and os.path.isdir(candidate):
-            return candidate
-
-    return results_dir_candidates[1]  # Fallback to default
 
 
 def _read_configs(configs_file):
